+++ src/scrapers/index.ts
@@ -1,4 +1,5 @@
 import { ScraperError } from "../types";
+import fetch from "node-fetch-extra";
 
 /**
  * Generate error object easily
@@ -11,6 +12,43 @@ export function error(status = -1, err = "Unknown"): ScraperError {
 		err,
 		success: false,
 	};
+}
+
+export async function getDataFromURL(url: string) {
+	let retryCount = 0;
+	let isValid = false;
+	let data: any = {};
+
+	while (!isValid && retryCount < 4) {
+		// Get data
+		const dataReq = await fetch(url);
+		if (dataReq.status === 204) {
+			// Empty result.
+			// Just end the loop
+			isValid = true;
+			return data;
+		} else {
+			let res = (data = await dataReq.text());
+			if (!res.startsWith("<") && res.trim().length > 0) {
+				try {
+					data = JSON.parse(data);
+					isValid = true;
+				} catch (e) {
+					// Oh well
+					retryCount++;
+					await sleep(100 * Math.floor(Math.random() * 50));
+				}
+			} else {
+				retryCount++;
+				await sleep(100 * Math.floor(Math.random() * 50));
+			}
+		}
+	}
+	return data;
+}
+
+function sleep(ms: number) {
+	return new Promise((resolve) => setTimeout(resolve, ms));
 }
 
 // Import Mangasee
