+++ src/scrapers/mangadex-v5.ts
@@ -1,6 +1,6 @@
 import chalk from "chalk";
 import fetch from "node-fetch-extra";
-import { error } from "./index";
+import { error, getDataFromURL } from "./index";
 import { Chapter, ScraperError, ScraperResponse } from "../types";
 import { Scraper, SearchOptions } from "./types";
 import { getProviderId, isProviderId } from "../routers/manga-page";
@@ -244,43 +244,6 @@ export class mangadexClass extends Scraper {
 			return error(-1, err);
 		}
 	}
-}
-
-async function getDataFromURL(url: string) {
-	let retryCount = 0;
-	let isValid = false;
-	let data: any = {};
-
-	while (!isValid && retryCount < 4) {
-		// Get data
-		const dataReq = await fetch(url);
-		if (dataReq.status === 204) {
-			// Empty result.
-			// Just end the loop
-			isValid = true;
-			return data;
-		} else {
-			let res = (data = await dataReq.text());
-			if (!res.startsWith("<") && res.trim().length > 0) {
-				try {
-					data = JSON.parse(data);
-					isValid = true;
-				} catch (e) {
-					// Oh well
-					retryCount++;
-					await sleep(100 * Math.floor(Math.random() * 50));
-				}
-			} else {
-				retryCount++;
-				await sleep(100 * Math.floor(Math.random() * 50));
-			}
-		}
-	}
-	return data;
-}
-
-function sleep(ms: number) {
-	return new Promise((resolve) => setTimeout(resolve, ms));
 }
 
 // Generate mangadex object and export it
